Detect branch from commit
